Also move librt.strings implementation
