+++ mypyc/lib-rt/strings/librt_strings.c
